Made Arm positions better
